+++ HISTORY.md
@@ -1,6 +1,15 @@
 # History
 
 ---
+## 0.5.0
+- HTML API:
+  - Added `text-search` attribute on `<li>` tag in order to send more context
+  to ONEm Platform to search in.
+
+- NodeJS API:
+  - Added `textSearch` attribute on both `MenuItem` and `FormItemMenuItem`.
+  This attribute has origin in `<li>` tag.
+---
 ## 0.4.0
 - HTML API:
   - Added `auto-select`, `multi-select` and `numbered` flags on `<section>`